requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
pyyaml==5.4.1
    # via -r requirements.in
